+++ src/hooks/useRankingItem.ts
@@ -1,6 +1,6 @@
 import { BASIC_ENDPOINT } from '@src/assets/endpoints';
 import type { Good } from '@src/types/Goods';
-import { useApiQuery } from '@src/hooks/useApiQuery';
+import { useApiQuery } from '@src/api/useApiQuery';
 
 export const useRankingItem = ({
   targetType,
